// src/Shopify/types/ShopifyOrderTypeCSV.ts
export type ShopifyOrderTypeCSV = {
  Name: string;
  Email: string;
  FinancialStatus: string;
  PaidAt: string;
  FulfillmentStatus: string;
  FulfilledAt: string;
  AcceptsMarketing: string;
  Currency: string;
  Subtotal: string;
  Shipping: string;
  Taxes: string;
  Total: string;
  DiscountCode: string;
  DiscountAmount: string;
  ShippingMethod: string;
  CreatedAt: string;
  LineitemQuantity: string;
  LineitemName: string;
  LineitemPrice: string;
  LineitemCompareAtPrice: string;
  LineitemSku: string;
  LineitemRequiresShipping: string;
  LineitemTaxable: string;
  LineitemFulfillmentStatus: string;
  BillingName: string;
  BillingStreet: string;
  BillingAddress1: string;
  BillingAddress2: string;
  BillingCompany: string;
  BillingCity: string;
  BillingZip: string;
  BillingProvince: string;
  BillingCountry: string;
  BillingPhone: string;
  ShippingName: string;
  ShippingStreet: string;
  ShippingAddress1: string;
  ShippingAddress2: string;
  ShippingCompany: string;
  ShippingCity: string;
  ShippingZip: string;
  ShippingProvince: string;
  ShippingCountry: string;
  ShippingPhone: string;
  Notes: string;
  NoteAttributes: string;
  CancelledAt: string;
  PaymentMethod: string;
  PaymentReference: string;
  RefundedAmount: string;
  Vendor: string;
  OutstandingBalance: string;
  Employee: string;
  Location: string;
  DeviceID: string;
  Id: string;
  Tags: string;
  RiskLevel: string;
  Source: string;
  LineitemDiscount: string;
  Phone: string;
  ReceiptNumber: string;
  BillingProvinceName: string;
  ShippingProvinceName: string;
  PaymentID: string;
  PaymentReferences: string;
};

export const ShopifyHeaderMap: Record<string, string> = {
  "Name": "Name",
  "Email": "Email",
  "Financial Status": "FinancialStatus",
  "Paid at": "PaidAt",
  "Fulfillment Status": "FulfillmentStatus",
  "Fulfilled at": "FulfilledAt",
  "Accepts Marketing": "AcceptsMarketing",
  "Currency": "Currency",
  "Subtotal": "Subtotal",
  "Shipping": "Shipping",
  "Taxes": "Taxes",
  "Total": "Total",
  "Discount Code": "DiscountCode",
  "Discount Amount": "DiscountAmount",
  "Shipping Method": "ShippingMethod",
  "Created at": "CreatedAt",
  "Lineitem quantity": "LineitemQuantity",
  "Lineitem name": "LineitemName",
  "Lineitem price": "LineitemPrice",
  "Lineitem compare at price": "LineitemCompareAtPrice",
  "Lineitem sku": "LineitemSku",
  "Lineitem requires shipping": "LineitemRequiresShipping",
  "Lineitem taxable": "LineitemTaxable",
  "Lineitem fulfillment status": "LineitemFulfillmentStatus",
  "Billing Name": "BillingName",
  "Billing Street": "BillingStreet",
  "Billing Address1": "BillingAddress1",
  "Billing Address2": "BillingAddress2",
  "Billing Company": "BillingCompany",
  "Billing City": "BillingCity",
  "Billing Zip": "BillingZip",
  "Billing Province": "BillingProvince",
  "Billing Country": "BillingCountry",
  "Billing Phone": "BillingPhone",
  "Shipping Name": "ShippingName",
  "Shipping Street": "ShippingStreet",
  "Shipping Address1": "ShippingAddress1",
  "Shipping Address2": "ShippingAddress2",
  "Shipping Company": "ShippingCompany",
  "Shipping City": "ShippingCity",
  "Shipping Zip": "ShippingZip",
  "Shipping Province": "ShippingProvince",
  "Shipping Country": "ShippingCountry",
  "Shipping Phone": "ShippingPhone",
  "Notes": "Notes",
  "Note Attributes": "NoteAttributes",
  "Cancelled at": "CancelledAt",
  "Payment Method": "PaymentMethod",
  "Payment Reference": "PaymentReference",
  "Refunded Amount": "RefundedAmount",
  "Vendor": "Vendor",
  "Outstanding Balance": "OutstandingBalance",
  "Employee": "Employee",
  "Location": "Location",
  "Device ID": "DeviceID",
  "Id": "Id",
  "Tags": "Tags",
  "Risk Level": "RiskLevel",
  "Source": "Source",
  "Lineitem discount": "LineitemDiscount",
  "Tax 1 Name": "Tax1Name",
  "Tax 1 Value": "Tax1Value",
  "Tax 2 Name": "Tax2Name",
  "Tax 2 Value": "Tax2Value",
  "Tax 3 Name": "Tax3Name",
  "Tax 3 Value": "Tax3Value",
  "Tax 4 Name": "Tax4Name",
  "Tax 4 Value": "Tax4Value",
  "Tax 5 Name": "Tax5Name",
  "Tax 5 Value": "Tax5Value",
  "Phone": "Phone",
  "Receipt Number": "ReceiptNumber",
  "Duties": "Duties",
  "Billing Province Name": "BillingProvinceName",
  "Shipping Province Name": "ShippingProvinceName",
  "Payment ID": "PaymentID",
  "Payment Terms Name": "PaymentTermsName",
  "Next Payment Due At": "NextPaymentDueAt",
  "Payment References": "PaymentReferences"
};